+++ packages/antd-management-fast-common/src/utils/constants.js
@@ -90,6 +90,7 @@ export const appInitDefault = {
   audioUploadMaxSize: 4,
   videoUploadMaxSize: 4,
   fileUploadMaxSize: 2,
+  tinymceScriptSrc: '',
   tinymceApiKey: '',
   tinymceImagesUploadUrl: '',
   layoutSetting: {