+++ packages/antd-management-fast-design-playground/src/components/displayer/DocumentPrintDesigner/DocumentContent/CellMarker/index.jsx
@@ -1,9 +1,13 @@
+import classNames from 'classnames';
 import React from 'react';
+
+import { isFunction } from 'easy-soft-utility';
 
 import styles from './index.less';
 
 function CellMarker(properties) {
   const {
+    useHover,
     data,
     highlight,
     highlightMode,
@@ -12,7 +16,10 @@ function CellMarker(properties) {
 
   return (
     <div
-      className={styles.cellMarker}
+      className={classNames(
+        styles.cellMarker,
+        useHover ? styles.cellMarkerEdit : null,
+      )}
       style={{
         position: 'absolute',
         width: '100%',
@@ -20,10 +27,18 @@ function CellMarker(properties) {
         ...(highlight ? { backgroundColor: 'rgba(90, 72, 0, 0.6)' } : {}),
       }}
       onClick={() => {
+        if (!isFunction(onClickCallback)) {
+          return;
+        }
+
         onClickCallback(data, highlightMode);
       }}
     ></div>
   );
 }
+
+CellMarker.defaultProps = {
+  useHover: true,
+};
 
 export { CellMarker };