+++ packages/antd-management-fast-design-playground/src/components/displayer/DocumentPrintDesigner/DocumentContent/Items/index.jsx
@@ -1,0 +1,2 @@
+export * from './LineItem';
+export * from './LineRemark';